% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-lm.R
\name{combine_tidy_estimates}
\alias{combine_tidy_estimates}
\title{Combine Tidy Estimates}
\usage{
combine_tidy_estimates(x, digits)
}
\arguments{
\item{x}{tibble. Expected to be \code{tidy.lm} output}

\item{digits}{integer. Significant figures to round estimates to.}
}
\description{
Combines the three \code{estimate}, \code{conf.low} and \code{conf.high} columns into a
"\code{estimate} (\code{conf.low} to \code{conf.high})" column.
}
\keyword{internal}
