% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{figures}
\alias{figures}
\alias{fig1_1_sample_sizes_by_region}
\title{Figures used in Environmental Impact of Housing in England}
\usage{
fig1_1_sample_sizes_by_region()
}
\description{
Methods that modularise each figure used in the final report.
}
