% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-reports.R
\name{get_knit_hook_source_hide}
\alias{get_knit_hook_source_hide}
\title{Knitr Hook to Hide Source Code}
\usage{
get_knit_hook_source_hide()
}
\description{
Creates a \code{source} hook to hide source code that could be of interest but
not pertinent to the report.
}
\details{
This returns a function that should be set using \code{knit_hook$set()}.
That is, \code{knitr::knit_hooks$set(source = get_knit_hook_source_hide())}.
}
