% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{to_label}
\alias{to_label}
\alias{LAD20TYPE_to_label}
\title{To Label}
\usage{
LAD20TYPE_to_label(x)
}
\arguments{
\item{x}{character. Coded data}
}
\description{
Convert coded data to a human-readable label
}
\section{Functions}{
\itemize{
\item \code{LAD20TYPE_to_label}: Local Authority types to label.
}}

