% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-imports.R
\name{utils-imports}
\alias{utils-imports}
\alias{where}
\title{Imports}
\usage{
where(fn)
}
\description{
Imports from other packages used within this package
}
\section{Links}{

\itemize{
\item \code{\link[rlang]{tidyeval-data}}, data pronoun.
\item \code{\link[rlang]{op-null-default}}, default value for NULL operator.
\item \code{\link[magrittr]{pipe}}, magrittr pipe (\verb{\%>\%})
\item \code{\link[tidyselect]{where}}, tidyselect keyword.
}
}

\keyword{internal}
