% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_final.R
\name{vis_final}
\alias{vis_final}
\alias{vis_sample_sizes_by_region}
\title{Visualisations}
\usage{
vis_sample_sizes_by_region()
}
\description{
Methods that take few arguments yet modularise each visualisation to be used
in the final report.
}
