% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{display}
\alias{display}
\alias{display_gorEHS}
\alias{display_YEAR}
\title{Display Version(s) of Data}
\usage{
display_gorEHS(x)

display_YEAR(x)
}
\arguments{
\item{x}{vector. Data to be converted.}
}
\description{
These functions convert data for understanding to a more aesthetically
pleasing version for visualisation.
}
\section{Functions}{
\itemize{
\item \code{display_gorEHS}: Line splits long region name, ``Yorkshire and ...`

\item \code{display_YEAR}: Assumes financial year so \code{2018} becomes \code{2018-19}.
}}

