% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{figures}
\alias{figures}
\alias{fig1_1_sample_sizes_by_region}
\alias{fig1_2_percentage_of_ownership}
\alias{fig1_3_fuelx_trend}
\alias{fig1_4_boiler_trend}
\alias{tbl1_1_percentage_of_flats}
\alias{tbl1_2_counts_of_wallinsx}
\alias{tbl2_1_linear_models}
\alias{fig3_1_population_effects}
\alias{fig3_2_population_effects}
\alias{tbl3_1_sigma}
\title{Figures used in Environmental Impact of Housing in England}
\usage{
fig1_1_sample_sizes_by_region()

fig1_2_percentage_of_ownership()

fig1_3_fuelx_trend()

fig1_4_boiler_trend()

tbl1_1_percentage_of_flats()

tbl1_2_counts_of_wallinsx()

tbl2_1_linear_models(fit_sap12, fit_EPceir12e)

fig3_1_population_effects(fit_mlm)

fig3_2_population_effects(fit_mlm)

tbl3_1_sigma(fit_mlm)
}
\arguments{
\item{fit_sap12, fit_EPceir12e}{Linear model objects.}

\item{fit_mlm}{Multilevel objects.}
}
\description{
Methods that modularise each figure used in the final report.
}
