% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{get_colour_scheme}
\alias{get_colour_scheme}
\alias{get_colour_scheme_region}
\title{Colour Schemes}
\usage{
get_colour_scheme(name)

get_colour_scheme_region()
}
\arguments{
\item{name}{choice. Must be a defined scheme.}
}
\description{
Consistency with colour schemes enforced by defining them here. Returns a
character vector as opposed to do something clever like RColorBrewer.
}
\section{Functions}{
\itemize{
\item \code{get_colour_scheme_region}: Wrapper to be used for region colouring (n=9).
}}

\section{Available Colour Scheme(s)}{

\itemize{
\item \code{ibm}: IBM Colour Blind safe scheme, nominal, 5 colours,
\url{https://lospec.com/palette-list/ibm-color-blind-safe}.
\item \code{moor64green} and \code{moor64blue}. Derived from MOOR64, ordinal, 6 colours,
\url{https://lospec.com/palette-list/moor64}.
\item \code{adobeseqforest} Adobe Sequential Forest, ordinal, 16 colours.
\url{https://spectrum.adobe.com/page/color-for-data-visualization}
\item \code{adobe12} Adobe Categorical, 12 colours.
\url{https://spectrum.adobe.com/page/color-for-data-visualization}
}
}

