% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{get_colour_scheme}
\alias{get_colour_scheme}
\title{Colour Schemes}
\usage{
get_colour_scheme(type)
}
\arguments{
\item{type}{choice. Must be a defined scheme.}
}
\description{
Consistency with colour schemes enforced by defining them here.
}
\section{Type(s)}{

- \code{YEAR}. Data can be 2014-15 to 2019-20
}

