% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAHS}
\alias{LAHS}
\title{Subset of Local Authority Housing Statistics}
\format{
A data frame with 3239 rows and ? variables:
\describe{
\item{LA_CODE}{Local authority code at the time the data was collected}
\item{LA_NAME}{Local Authority name at the time the data was collected}
\item{LA_TYPE}{Current local authority type (April 2020)}
\item{LA_REGION}{English Region name}
\item{YEAR}{The starting year of the financial reporting year}
}
}
\usage{
LAHS
}
\description{
We focus on the subset of the data that could be relevant from an
environmental perspective.
}
\keyword{datasets}
